Make CircleCI config consistent with Staff SSO
The previous commit updated the runtime versions for development
purposes, this commit updates the runtimes in CI. Adjacent to this
change was a general update to bring the CI config in line with
staff-sso.

This has allowed us to remove all but one of the requirements file. More
work could be done in the future to properly seperate the dependencies
for running and testing the application.