Remove dependency on `py` library
The `pytest-sugar` version 0.9.4 is dependant on the `py` library, by
upgrading it to 0.9.7 we no longer require the `py` library.

Removing this dependancy future-proofs the codebase since the `py`
library is no longer maintained and only supports upto Python 3.10,
which we are currently using. It also resolves a high severity dependabot alert, though we don't invoke the vulnerable code.